requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-gis==0.18
eunjeon==0.4.0
future==0.18.2
idna==3.3
oauthlib==3.2.0
psycopg2==2.9.3
python-twitter==3.5
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
